requirements.txt
flask==3.1.0
requests==2.32.3
markdown==3.7
anthropic>=0.49.0
openai==1.60.0
google-genai==1.67.0
python-dotenv==1.0.1
